feat: structure optimization
