+++ packages/playground/src/components/filters/filters-bar/filter-variants/calendar.tsx
@@ -1,20 +1,20 @@
 import { useState } from 'react'
-import { Calendar } from '@harnessio/canary'
-import { FilterValue } from '../types'
+import { Calendar as UICalendar } from '@harnessio/canary'
+import { FilterValue } from '../../types'
 
-interface DateFilterProps {
+interface CalendarProps {
   filter: FilterValue
   onUpdateFilter: (type: string, selectedValues: string[]) => void
 }
 
-const DateFilter = ({ filter, onUpdateFilter }: DateFilterProps) => {
+const Calendar = ({ filter, onUpdateFilter }: CalendarProps) => {
   const [date, setDate] = useState<Date | undefined>(
     filter.selectedValues[0] ? new Date(filter.selectedValues[0]) : undefined
   )
 
   return (
     <div className="mt-2.5 px-2 pb-3 pt-1">
-      <Calendar
+      <UICalendar
         className="p-0"
         mode="single"
         selected={date}
@@ -30,4 +30,4 @@ const DateFilter = ({ filter, onUpdateFilter }: DateFilterProps) => {
   )
 }
 
-export default DateFilter
+export default Calendar
