+++ packages/playground/src/components/filters/index.ts
@@ -1,9 +1,9 @@
 import Filters from './filters'
-import FiltersSelectedBar from './filters-selected-bar'
+import FiltersBar from './filters-bar/filters-bar'
 import useFilters from './use-filters'
 
 import { FilterCondition, FilterOption, SortDirection, SortOption } from './types'
 
-export { Filters, FiltersSelectedBar, useFilters }
+export { Filters, FiltersBar, useFilters }
 
 export type { FilterOption, FilterCondition, SortOption, SortDirection }
